requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
ghp-import==2.1.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.3
mergedeep==1.3.4
mkdocs==1.4.3
mkdocs-build-plantuml-plugin==1.7.4
mkdocs-gen-files==0.5.0
mkdocs-htmlproofer-plugin==0.13.1
mkdocs-material==9.1.18
mkdocs-material-extensions==1.1.1
mkdocs-meta-descriptions-plugin==2.2.0
mkdocs-section-index==0.3.5
packaging==23.1
Pygments==2.15.1
pymdown-extensions==10.1
pyparsing==3.1.0
python-dateutil==2.8.2
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2023.6.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
urllib3==2.0.3
watchdog==3.0.0
